requirements.txt
flask
supabase-py
gunicorn
sqlalchemy
flask_sqlalchemy
psycopg2-binary
python-dotenv 